Remove unneeded includes in header files.
