+++ searchsummary/src/vespa/searchsummary/docsummary/CMakeLists.txt
@@ -6,13 +6,15 @@ vespa_add_library(searchsummary_docsummary OBJECT
     attribute_field_writer.cpp
     attributedfw.cpp
     check_undefined_value_visitor.cpp
+    copy_dfw.cpp
     docsumconfig.cpp
-    docsumfieldwriter.cpp
-    docsumstate.cpp
+    docsum_field_writer.cpp
     docsum_store_document.cpp
+    docsumstate.cpp
     docsumstorevalue.cpp
     docsumwriter.cpp
     dynamicteaserdfw.cpp
+    empty_dfw.cpp
     general_result.cpp
     geoposdfw.cpp
     getdocsumargs.cpp
@@ -22,10 +24,12 @@ vespa_add_library(searchsummary_docsummary OBJECT
     matched_elements_filter_dfw.cpp
     positionsdfw.cpp
     rankfeaturesdfw.cpp
+    res_type_utils.cpp
     resultclass.cpp
     resultconfig.cpp
     resultpacker.cpp
     searchdatatype.cpp
+    simple_dfw.cpp
     struct_fields_resolver.cpp
     struct_map_attribute_combiner_dfw.cpp
     summaryfeaturesdfw.cpp
