+++ searchsummary/src/vespa/searchsummary/docsummary/array_attribute_combiner_dfw.h
@@ -3,6 +3,7 @@
 #pragma once
 
 #include "attribute_combiner_dfw.h"
+#include <vector>
 
 namespace search::attribute { class IAttributeContext; }
