+++ searchsummary/src/vespa/searchsummary/docsummary/attribute_combiner_dfw.h
@@ -2,7 +2,8 @@
 
 #pragma once
 
-#include "docsumfieldwriter.h"
+#include "simple_dfw.h"
+#include <memory>
 
 namespace search {
 class MatchingElements;
@@ -21,7 +22,7 @@ class DynamicDocsumWriter;
  * This class reads values from multiple struct field attributes and
  * inserts them as an array of struct or a map of struct.
  */
-class AttributeCombinerDFW : public ISimpleDFW
+class AttributeCombinerDFW : public SimpleDFW
 {
 protected:
     uint32_t _stateIndex;
@@ -36,8 +37,8 @@ public:
     ~AttributeCombinerDFW() override;
     bool IsGenerated() const override;
     bool setFieldWriterStateIndex(uint32_t fieldWriterStateIndex) override;
-    static std::unique_ptr<IDocsumFieldWriter> create(const vespalib::string &fieldName, search::attribute::IAttributeContext &attrCtx,
-                                                      bool filter_elements, std::shared_ptr<MatchingElementsFields> matching_elems_fields);
+    static std::unique_ptr<DocsumFieldWriter> create(const vespalib::string &fieldName, search::attribute::IAttributeContext &attrCtx,
+                                                     bool filter_elements, std::shared_ptr<MatchingElementsFields> matching_elems_fields);
     void insertField(uint32_t docid, GetDocsumsState *state, ResType type, vespalib::slime::Inserter &target) override;
 };
