+++ searchsummary/src/vespa/searchsummary/docsummary/attributedfw.h
@@ -2,38 +2,38 @@
 
 #pragma once
 
-#include "docsumfieldwriter.h"
-#include "docsumstate.h"
+#include "simple_dfw.h"
+#include <memory>
 
-namespace search { class MatchingElementsFields; }
+namespace search {
+class IAttributeManager;
+class MatchingElementsFields;
+}
 namespace search::attribute { class IAttributeVector; }
 
 namespace search::docsummary {
 
 /**
- * Factory to create an IDocsumFieldWriter to write an attribute vector to slime.
+ * Factory to create an DocsumFieldWriter to write an attribute vector to slime.
  */
 class AttributeDFWFactory {
 public:
-    static std::unique_ptr<IDocsumFieldWriter> create(IAttributeManager& attr_mgr,
-                                                      const vespalib::string& attr_name,
-                                                      bool filter_elements = false,
-                                                      std::shared_ptr<MatchingElementsFields> matching_elems_fields
-                                                      = std::shared_ptr<MatchingElementsFields>());
+    static std::unique_ptr<DocsumFieldWriter> create(IAttributeManager& attr_mgr,
+                                                     const vespalib::string& attr_name,
+                                                     bool filter_elements = false,
+                                                     std::shared_ptr<MatchingElementsFields> matching_elems_fields = std::shared_ptr<MatchingElementsFields>());
 };
 
-class AttrDFW : public ISimpleDFW
+class AttrDFW : public SimpleDFW
 {
 private:
     vespalib::string _attrName;
 protected:
-    const attribute::IAttributeVector& get_attribute(const GetDocsumsState& s) const {
-        return *s.getAttribute(getIndex());
-    }
-    const vespalib::string & getAttributeName() const override { return _attrName; }
+    const attribute::IAttributeVector& get_attribute(const GetDocsumsState& s) const;
+    const vespalib::string & getAttributeName() const override;
 public:
     AttrDFW(const vespalib::string & attrName);
-    bool IsGenerated() const override { return true; }
+    bool IsGenerated() const override;
 };
 
 }
