+++ searchsummary/src/vespa/searchsummary/docsummary/geoposdfw.cpp
@@ -6,6 +6,7 @@
 #include <vespa/searchlib/common/location.h>
 #include <vespa/vespalib/util/jsonwriter.h>
 #include <vespa/vespalib/data/slime/cursor.h>
+#include <vespa/vespalib/data/slime/inserter.h>
 #include <vespa/vespalib/stllike/asciistream.h>
 #include <vespa/vespalib/util/issue.h>
 #include <climits>
