+++ searchsummary/src/vespa/searchsummary/docsummary/positionsdfw.cpp
@@ -270,8 +270,9 @@ PositionsDFW::UP PositionsDFW::create(const char *attribute_name, IAttributeMana
     return std::make_unique<PositionsDFW>(attribute_name, useV8geoPositions);
 }
 
-AbsDistanceDFW::UP AbsDistanceDFW::create(const char *attribute_name, IAttributeManager *attribute_manager) {
-    AbsDistanceDFW::UP ret;
+std::unique_ptr<DocsumFieldWriter>
+AbsDistanceDFW::create(const char *attribute_name, IAttributeManager *attribute_manager) {
+    std::unique_ptr<DocsumFieldWriter> ret;
     if (attribute_manager != nullptr) {
         if (!attribute_name) {
             LOG(debug, "createAbsDistanceDFW: missing attribute name '%p'", attribute_name);
