+++ searchsummary/src/vespa/searchsummary/docsummary/resultpacker.cpp
@@ -1,6 +1,7 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include "resultpacker.h"
+#include "resultconfig.h"
 #include <vespa/searchcommon/common/undefinedvalues.h>
 #include <vespa/vespalib/util/size_literals.h>
