+++ searchsummary/src/vespa/searchsummary/docsummary/summaryfeaturesdfw.cpp
@@ -3,6 +3,7 @@
 #include "summaryfeaturesdfw.h"
 #include "docsumstate.h"
 #include <vespa/vespalib/data/slime/cursor.h>
+#include <vespa/vespalib/data/slime/inserter.h>
 
 #include <vespa/log/log.h>
 LOG_SETUP(".searchlib.docsummary.summaryfeaturesdfw");
