+++ streamingvisitors/src/vespa/vsm/vsm/docsumconfig.cpp
@@ -1,14 +1,16 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include <vespa/vsm/vsm/docsumconfig.h>
-#include <vespa/searchsummary/docsummary/docsumfieldwriter.h>
+#include <vespa/searchsummary/docsummary/copy_dfw.h>
+#include <vespa/searchsummary/docsummary/empty_dfw.h>
 #include <vespa/searchsummary/docsummary/matched_elements_filter_dfw.h>
+#include <vespa/searchsummary/docsummary/resultconfig.h>
 #include <vespa/searchlib/common/matching_elements_fields.h>
 #include <vespa/vsm/config/config-vsmfields.h>
 #include <vespa/vsm/config/config-vsmsummary.h>
 
 using search::MatchingElementsFields;
-using search::docsummary::IDocsumFieldWriter;
+using search::docsummary::DocsumFieldWriter;
 using search::docsummary::CopyDFW;
 using search::docsummary::EmptyDFW;
 using search::docsummary::MatchedElementsFilterDFW;
@@ -41,10 +43,10 @@ DynamicDocsumConfig::DynamicDocsumConfig(search::docsummary::IDocsumEnvironment*
 {
 }
 
-IDocsumFieldWriter::UP
+std::unique_ptr<DocsumFieldWriter>
 DynamicDocsumConfig::createFieldWriter(const string & fieldName, const string & overrideName, const string & argument, bool & rc, std::shared_ptr<search::MatchingElementsFields> matching_elems_fields)
 {
-    IDocsumFieldWriter::UP fieldWriter;
+    std::unique_ptr<DocsumFieldWriter> fieldWriter;
     if ((overrideName == "staticrank") ||
         (overrideName == "ranklog") ||
         (overrideName == "label") ||
